requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.2
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
boto3==1.35.0
botocore==1.35.0
openai==1.40.0
pdfplumber==0.11.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.23.8
playwright==1.46.0
PyYAML==6.0.2
psycopg2-binary>=2.9
pykwalify>=1.8
pytest-cov>=4.1
